Pure refactor of classifiers and place detection
